requirements.txt
gspread
google-auth
requests
python-dotenv
pytz